Fix rendering of bootstrapped client
When loading the home page, the "bootstrapped client" block would always
show initializing despite that we are providing it with the correct
feature flag data on instantiation.

Because `<script>` tags block rendering of the DOM, and because a
bootstrapped client is immediately initialized, when we try to update
the text for the appropriate LI tag, we find that it doesn't exist in
the DOM yet.

This wasn't an issue for the standard JS client because it has to
actually connect to LD and fetch the flags asynchronously. This allowed
processing to continue and the DOM to finish rendering.

By moving JS execution to the bottom of the page, we can ensure the tags
are present as the time of initialization.