Fix rendering of bootstrapped client (#7)
When loading the home page, the "bootstrapped client" block would 
show initializing even though we are bootstrapping it on instantiation.

Because `<script>` tags block rendering of the DOM, and because a
bootstrapped client is immediately initialized, when we try to update
the text for the appropriate LI tag, we find that it doesn't exist in
the DOM yet.

This wasn't an issue for the standard JS client because it has to
actually connect to LD and fetch the flags asynchronously. This allowed
processing to continue and the DOM to finish rendering.

To resolve this, we wait until the DOM has fully loaded before
attempting to run any of our initialization code.